Seguridad ajax + configuracion imagen perfil

+++ temas/original/ajax/configuracion_perfil_imagen.php
@@ -1,5 +1,5 @@
 <?php		
-	if ( isset($_SESSION['username'])) {
+	if (!empty($_SERVER['HTTP_X_REQUESTED_WITH']) && strtolower($_SERVER['HTTP_X_REQUESTED_WITH']) == 'xmlhttprequest' && isset($_SESSION['username'])) {
 		$ruta = 'static-content/perfiles/';
 
 		if (isset($_FILES['file_0'])) {
@@ -25,14 +25,10 @@
 		echo '<strong>Imagen no válida</strong>. Solo fotos con extenciones jpeg y png. Tu extencion es: '.$finfo;
 	} else if($ancho != $alto) {
 		echo "Solo se aceptan <strong>imagenes cuadradas!</strong>. Tu imagen tiene un tamaño de: <strong>".$ancho."x".$alto."</strong>";
-	} else if($imagen['size'] > 3000000) {
+	} else if($imagen['size'] > 3145728) {
 		echo 'Foto invalida, la imagen exede los 3MB!';
 	} elseif($imagen['error'] > 0) {
 		echo 'Error al subir la imagen!';
-	/*} elseif (($_FILES['archivo']['type'] != 'image/*') and ($_FILES['archivo']['type'] = 'image/gif')) {
-		echo 'Solo puede subir imagenes exeptuando los gifs!';*/
-	} elseif(file_exists($ruta.$name)) {
-		echo 'Error anormal, reporte. Intente nuevamente.';
 	
 	/*
 	----------------
@@ -43,13 +39,13 @@
 	} else {
 		$select_img_pf_o = $db->query("SELECT imagen_perfil FROM cuentas WHERE idcuenta = ".$pf['idcuenta']);
 		$select_img_pf = $select_img_pf_o->fetch_array();
-		unlink($ruta.$select_img_pf['imagen_perfil']);
-		
+		if (file_exists($ruta.$select_img_pf['imagen_perfil'])) {
+			unlink($ruta.$select_img_pf['imagen_perfil']);
+		}
 		$cambiar_img_pf = $db->query("UPDATE cuentas SET imagen_perfil = '".$name."' WHERE idcuenta = ".$pf['idcuenta']);
 		move_uploaded_file($imagen['tmp_name'], $ruta.$name);
 		echo "Finalizado";
-
 		}
 	} else {
-		echo 'No tienes permiso para entrar aqui!';
+		header('Location: ?p=404');
 		}
